Cleaned up, moved font atlas functions to font module
